+++ doc/changelog.rst
@@ -6,6 +6,7 @@ Changelog
 
 Fixed
 ^^^^^
+- Do not raise exceptions when encountering SCIM errors. #3
 - Invalid domains and network errors are properly handled. #6
 
 [0.1.3] - 2024-07-25