+++ doc/changelog.rst
@@ -1,6 +1,9 @@
 Changelog
 =========
 
+[0.1.5] - 2024-09-01
+--------------------
+
 Fixed
 ^^^^^
 - check_random_url error after scim2-client 0.2.0 update. #8